Move detect_encoding function and column extraction codes to utils
